sklearn-project/requirements.txt
scikit-learn==0.23.1
joblib==0.15.1
neptune-client==0.4.119
